added a Loading component to home and menu page
