+++ frontend/src/components/common/Loading/Loading.jsx
@@ -1,0 +1,9 @@
+import { LoadingContainer } from "@/styles/components/common/Loading/LoadingStyle";
+
+export function Loading() {
+	return (
+		<LoadingContainer>
+			<h2>Loading....</h2>
+		</LoadingContainer>
+	);
+}
