+++ frontend/src/styles/components/common/Loading/LoadingStyle.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const LoadingContainer = styled.div`
+	margin: 60px;
+
+	h2 {
+		text-align: center;
+
+		@media (max-width: 520px) {
+			font-size: 0.8rem;
+		}
+	}
+`;
